requirements.txt
moviepy
librosa
matplotlib
torch
torchvision
Pillow
pandas
timm